+++ src/gruppe22/cdio/business/IBusinessLogic.java
@@ -17,7 +17,7 @@ public interface IBusinessLogic {
     int getUserId();
     List<String> getUserList();
     UserDTO getUser(int userId) throws IUserDAO.DALException;
-    void createUser(int userId, String userName, String ini, int cpr, String role);
-    void updateUser(int userId, String userName, int cprNumber, String ini, String pw, List<String> roles) throws IUserDAO.DALException;
+    void createUser(int userId, String userName, String ini, long cpr, String role);
+    void updateUser(int userId, String userName, long cprNumber, String ini, String pw, List<String> roles) throws IUserDAO.DALException;
     void deleteUser(int userId);
 }
